+++ 01-Search/src/sort.c
@@ -4,7 +4,7 @@
 #include "sort.h"
 
 void merge(int *arr, int left, int mid, int right, int *tmp) {
-  size_t i=0, j=0;
+  int i=0, j=0;
 
   while (left+i < mid && mid+j < right) {
     if (*(arr + left+i) < *(arr + mid+j)) {
@@ -28,9 +28,9 @@ void merge_sort(int* arr, int n) {
   int *tmp = (int*)malloc(n * sizeof(int));
   if (!tmp) return;
   
-  size_t min=0;
-  for (size_t i=1; i<n; i*=2)
-    for (size_t j=0; j < n-i; j += 2*i) {
+  int min=0;
+  for (int i=1; i<n; i*=2)
+    for (int j=0; j < n-i; j += 2*i) {
       min = (j+2*i < n) ? j+2*i : n;
       merge(arr, j, i+j, min, tmp);
     }
