+++ libs/utils/src/utils.c
@@ -1,6 +1,6 @@
 #include <sys/time.h>
 #include <stdlib.h>
-#include <utils/utils.h>
+#include <utils.h>
 
 // current time
 double wtime() {
@@ -11,12 +11,12 @@ double wtime() {
 
 // generate random number
 int getrand(int min, int max) {
-  return rand() / (RAND_MAX + 1) * (max - min) + min;
+  return (double)rand() / (RAND_MAX + 1.0) * (max - min) + min;
 }
 
 // fill the array with random numbers
-int* fillarr(int* arr, size_t n, int min, int max) {
+int* fillarr(int* arr, int n, int min, int max) {
   if (n==0) return NULL;
-  for (size_t i=0; i<n; i++) *(arr+i) = getrand(min, max);
+  for (int i=0; i<n; i++) *(arr+i) = getrand(min, max);
   return arr;
 }
